add in getAllReferencedIds
