+++ packages/adapter-components/index.ts
@@ -28,3 +28,4 @@ export * from './src/references/element_fixers'
 export * from './src/references/custom_references'
 export * from './src/references/weak_reference_handler'
 export * from './src/resolve_utils'
+export * from './src/references/get_references'
