+++ packages/adapter-components/test/references/get_references.test.ts
@@ -1,0 +1,30 @@
+/*
+*                      Copyright 2024 Salto Labs Ltd.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with
+* the License.  You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+import { getAllReferencedIds } from '../../src/references/get_references'
+import { mockInstance, templateElemID, templateElemID2 } from '../utils'
+
+
+describe('getAllReferencedIds', () => {
+  it('should find referenced ids', () => {
+    const res = getAllReferencedIds(mockInstance)
+    expect(res).toEqual(new Set(['mockAdapter.test', 'mockAdapter.test2.field.aaa',
+      templateElemID.getFullName(), templateElemID2.getFullName()]))
+  })
+  it('should find referenced ids only in annotations', () => {
+    const res = getAllReferencedIds(mockInstance, true)
+    expect(res).toEqual(new Set(['mockAdapter.test']))
+  })
+})
